Extract Annotations to own file
