+++ src/annotations/Annotation.ts
@@ -1,0 +1,5 @@
+export type AnnotationConstructor = ObjectConstructor & { _name: string; };
+
+export abstract class Annotation {
+  public static _name: string;
+}
